move and rename some files
