+++ src/sim/sim_adapter.cpp
@@ -1,4 +1,4 @@
-#include "backend_adapter.hpp"
+#include "sim_adapter.hpp"
 
 ftb_slot _ftb_global_slots[FTBACKEND_SLOTS];
